Hint that map-instructions hash tables grow fast
map-instructions-xxx functions use a hash table to keep
track of instructions that have been seen already.
These hash tables grow quickly and then are discarded.
I added a :size 1024 to start them out large and :rehash-size 4.0
so that they grow by a multiplicative factor of 4.0 when they are
rehashed. This leads to less rehashing of these heavily used hash tables.